refactor: HttpCookie, HttpHeader 기본 생성자 -> 팩토리 메서드 패턴
